refactor: split alerts.rs into focused submodules [OPE-487]
Transform the monolithic alerts.rs (422 LOC) into a modular directory
structure with focused submodules:

- types.rs: AlertDispatchError and DEFAULT_COOLDOWN constant
- dispatcher.rs: AlertDispatcher struct and all evaluation/action logic
- tests.rs: all 8 unit tests
- mod.rs: re-exports public API
